// src/types/hr-management/employee-management.d.ts
export interface ISelect {
    minGrade?: number;
    maxGrade?: number;
    id?: string;
    label?: number | string;
    value?: number | string;
}